refactor(workflows): neutralize the automation access service filename
