+++ turbo/apps/api/src/signals/services/workflow-webhook-automation.service.ts
@@ -32,7 +32,7 @@ import type {
   AutomationRow,
 } from "./zero-workflow-automation-launch.service";
 import type { WorkflowAutomationContext } from "./workflow-automation-context.service";
-import { workflowAutomationCanFire } from "./zero-workflow-automation-access.service";
+import { workflowAutomationCanFire } from "./workflow-automation-access.service";
 import { ensureWorkflowUserAutomationThread } from "./zero-workflow-user-automation-thread.service";
 import { loadOrgPlanCapabilities } from "./org-plan-entitlement-read.service";
